Created userProfileUtils to format userProfiles returned by the backend Added acceptCandidateToTeam method to the team controller to add a candidate as a team member Fixed issue on the team model where the team code would be regenerated when updates are saved in a team Added new route to team routes to handle accepting a candidate to a team Modified CandidateCard to handle when team members accept a candidate to join their team Modified Participant ProfileInfo to avoid errors when the user data passed is null or undefined Modified Participant ProfileSide, ProfileTop and Profile to handled cases when prop data is passed as null or undefined Created RecruitementFavorites component to separate favorites from the Profile component Modified objToArr inside dataModifiers utility to handle cases when the object passed is not an object type Modified candidates page to render data on the same page conditionally based on the state Modified profile team page to hide team profile when the user is editing or creating a team Added acceptCandidateToTeam method to the dashboard actions Added acceptCandidateToTeam method to the teams services

+++ backend/modules/team/controller.js
@@ -120,6 +120,32 @@ controller.joinTeam = (eventId, userId, code) => {
     })
 }
 
+controller.acceptCandidateToTeam = (eventId, userId, code, candidateId) => {
+    console.log('eventId', eventId)
+    console.log('userId', userId)
+    console.log('code', code)
+    console.log('candidateId', candidateId)
+    return controller.getTeamByCode(eventId, code).then(team => {
+        console.log('is user owner', team.owner === userId)
+        console.log('is user member', _.includes(team.members, userId))
+        console.log(
+            'can it proceed?',
+            team.owner !== userId || !_.includes(team.members, userId),
+        )
+        if (!_.includes([team.owner].concat(team.members), userId)) {
+            throw new InsufficientPrivilegesError(
+                'Only the team owner can accept candidates',
+            )
+        }
+        team.members = team.members.concat(candidateId)
+        team.candidates = team.candidates.filter(
+            candidate => candidate.userId !== candidateId,
+        )
+        console.log('team after action on controller', team)
+        return team.save()
+    })
+}
+
 controller.candidateApplyToTeam = (eventId, userId, code, applicationData) => {
     // const team = controller.getTeam(eventId, applicationData.teamId)
     return controller.getTeamByCode(eventId, code).then(team => {
@@ -149,7 +175,7 @@ controller.candidateApplyToTeam = (eventId, userId, code, applicationData) => {
 controller.leaveTeam = (eventId, userId) => {
     return controller.getTeam(eventId, userId).then(team => {
         team.members = team.members.filter(member => member !== userId)
-        if (team.members.length === 0) {
+        if (team.members.length === 0 && team.owner === userId) {
             controller.deleteTeam(eventId, userId)
         } else {
             return team.save()
